Fix copy title card background color in x/p chain
