+++ src/app/pages/CChainPages/Address/index.tsx
@@ -1,5 +1,5 @@
 import React, { FC } from 'react';
-import { Paper, Box, useTheme } from '@mui/material';
+import { Paper } from '@mui/material';
 import { mdiFileDocumentOutline } from '@mdi/js';
 import PageContainer from 'app/components/PageContainer';
 import TabsHeader from 'app/components/TabComponent/TabsHeader';
@@ -8,8 +8,7 @@ import Transactions from './Transactions';
 import SubPageTitle from 'app/components/SubPageTitle';
 import useWidth from 'app/hooks/useWidth';
 import { CCHAIN } from 'utils/route-paths';
-import DetailsField from 'app/components/DetailsField';
-import Icon from '@mdi/react';
+import CopyTitleCard from 'app/components/CopyTitleCard';
 
 const tabOptions = [
   {
@@ -20,7 +19,6 @@ const tabOptions = [
 
 const CAddressDetails: FC = () => {
   const { isDesktop } = useWidth();
-  const theme = useTheme();
   const [value, setValue] = React.useState(0);
   const address = window.location.pathname.split('/').pop() as string;
 
@@ -33,32 +31,12 @@ const CAddressDetails: FC = () => {
       metaContent="Address Detail C-Chain"
     >
       <SubPageTitle title="Address Detail" backToLink={CCHAIN} />
-      <Box
-        sx={{
-          backgroundColor: 'primary.dark',
-          border: theme.palette.mode === 'dark' ? 'solid 1px' : '0px',
-          borderColor: 'borders.main',
-          borderRadius: '7px',
-          width: '100%',
-          display: 'flex',
-          alignItems: 'center',
-        }}
-      >
-        <DetailsField
-          field="Address"
-          value={address}
-          type="string"
-          icon={
-            <Icon
-              path={mdiFileDocumentOutline}
-              color="latestList.iconColor"
-              style={{ width: '20px', height: '20px' }}
-            />
-          }
-          allowCopy={true}
-          style={{ padding: '1rem' }}
-        />
-      </Box>
+      <CopyTitleCard
+        label="Address"
+        value={address}
+        icon={mdiFileDocumentOutline}
+        mixedStyle
+      />
       <Paper
         square
         variant="outlined"
